requirements.txt
# Core RAG Framework
langchain>=0.1.0
langchain-community>=0.0.20
langchain-huggingface>=0.0.1

# Embeddings and LLM
transformers>=4.35.0
sentence-transformers>=2.2.2
huggingface-hub>=0.17.0
torch>=2.0.0  # For running transformers; use torchvision torchaudio if needed
accelerate>=0.21.0  # For efficient inference

# Vector Storage
faiss-cpu>=1.7.4  # Use faiss-gpu for GPU support

# Data Processing and Parsing
pydantic>=2.0.0
pyyaml>=6.0
pandas>=2.0.0
polars>=0.19.0  # Alternative to pandas for faster processing

# AWS Integrations (optional for local dev, required for prod)
boto3>=1.28.0

# API and UI
fastapi>=0.104.0
uvicorn>=0.24.0  # ASGI server for FastAPI
streamlit>=1.28.0  # For chat UI

# Orchestration (if using local Airflow)
apache-airflow>=2.7.0  # Optional, for pipeline scheduling

# Testing and Utilities
pytest>=7.4.0
pytest-cov>=4.1.0
structlog>=23.1.0  # Structured logging

# Deployment
docker  # Not pip-installable, but note for setup

# Notes:
# - For GPU: pip install faiss-gpu torch --index-url https://download.pytorch.org/whl/cu118
# - Install with: pip install -r requirements.txt
# - This stack prioritizes open-source for banking compliance and cost-efficiency.